movie-tickets-backend/src/main/java/com/movietickets/model/Movie.java
package com.movietickets.model;

public class Movie {
	private String imdbid;
	private String title;
	private String genre;
	private String plot;
	private String poster;
	private String runtime;
	private String director;
	private String actors;
	private int year;
	private double defaultprice;
	
	
	public String getImdbid() {
		return imdbid;
	}
	public void setImdbid(String imdbid) {
		this.imdbid = imdbid;
	}
	public String getTitle() {
		return title;
	}
	public void setTitle(String title) {
		this.title = title;
	}
	public String getGenre() {
		return genre;
	}
	public void setGenre(String genre) {
		this.genre = genre;
	}
	public String getPlot() {
		return plot;
	}
	public void setPlot(String plot) {
		this.plot = plot;
	}
	public String getPoster() {
		return poster;
	}
	public void setPoster(String poster) {
		this.poster = poster;
	}
	public String getRuntime() {
		return runtime;
	}
	public void setRuntime(String runtime) {
		this.runtime = runtime;
	}
	public String getDirector() {
		return director;
	}
	public void setDirector(String director) {
		this.director = director;
	}
	public String getActors() {
		return actors;
	}
	public void setActors(String actors) {
		this.actors = actors;
	}
	public int getYear() {
		return year;
	}
	public void setYear(int year) {
		this.year = year;
	}
	public double getDefaultprice() {
		return defaultprice;
	}
	public void setDefaultprice(double defaultprice) {
		this.defaultprice = defaultprice;
	}
	
	
}